Armor Underwear, Thirst Config, juice bug fix, moved serene to compat
Armor Underwear compat works (ozzy, goopak, liner, just not antifreeze), thirst is configurable too.

Not ready for a release yet, there's some stuff I want to do first.

This'll break anyone's serene seasons' config unfortunately

+++ src/main/java/com/charles445/simpledifficulty/temperature/ModifierBlocksTiles.java
@@ -89,12 +89,12 @@ public class ModifierBlocksTiles extends ModifierBase
 		if(result > hottestValue)
 		{
 			//Hotter than hottestValue, clamp
-			return Math.min(hottestValue + ModConfig.server.temperature.stackingTemperatureLimit, result);
+			return Math.min(hottestValue + (float)ModConfig.server.temperature.stackingTemperatureLimit, result);
 		}
 		else if(result < coldestValue)
 		{
 			//Colder than coldestValue, clamp
-			return Math.max(coldestValue - ModConfig.server.temperature.stackingTemperatureLimit, result);
+			return Math.max(coldestValue - (float)ModConfig.server.temperature.stackingTemperatureLimit, result);
 		}
 		else
 		{
